+++ packages/card/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@1.11.5...@ant-design/pro-card@1.11.6) (2021-03-31)
+
+### Bug Fixes
+
+- **card:** remove antd charts demos, replaced with img ([#2343](https://github.com/ant-design/pro-components/issues/2343)) ([cb69b10](https://github.com/ant-design/pro-components/commit/cb69b104ba0be6ad94de118cb60a9e03834ba2ba))
+
 ## [1.11.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@1.11.4...@ant-design/pro-card@1.11.5) (2021-03-29)
 
 **Note:** Version bump only for package @ant-design/pro-card